testing staging deploy script

+++ deploy-staging.sh
@@ -8,9 +8,13 @@ if [ "$branch" = staging ]
 then
   if [ -z "$(git status --porcelain)" ]
   then
-    rm CNAME
-    git commit -am "remove CNAME"
-    git push staging staging
+    ls CNAME
+    if ls CNAME;
+    then
+      rm CNAME
+      git commit -am "remove CNAME"
+      git push cole_and_lopez_stagingng staging
+    fi
     echo your changes will soon appear at https://ensconced.github.io/cole_and_lopez_staging/
   else
     echo git tree must be clean
